update server and app to simplify server initializatio and DB connection
