+++ server.js
@@ -1,46 +1,30 @@
-import express from 'express';
-import bodyParser from 'body-parser';
-import cors from 'cors';
-import dotenv from "dotenv";
-import listEndpoints from 'express-list-endpoints';
-import { connectDatabase } from './config/database';
+import { connectDatabase } from "./config/database.js";
+import { seedDatabase } from "./utils/seedDatabase.js";
+import { app } from "./app.js";
 
-dotenv.config(); // Load environment variables from .env
-
-connectDatabase();  // Connect to MongoDB
-
-// Defines the port the app will run on. Defaults to 8080, but can be 
-// overridden when starting the server. For example:
-//
-// PORT=9000 npm start
 const port = process.env.PORT || 8070;
-const app = express();
 
-// Add middlewares to enable cors and json body parsing
-app.use(cors());
-app.use(bodyParser.json());
-app.use((req, res, next) => { 
-  if (mongoose.connection.readyState == 1) {
-    next()
-  } else {
-    res.status(503).json({ error: "Service unavailable"})
+(async () => {
+  try {
+    // Connect to MongoDB
+    await connectDatabase();
+
+    // Check if RESET_DATABASE is set to trigger seeding
+    if (process.env.RESET_DATABASE === "true") {
+      console.log("RESET_DATABASE is set. Seeding the database...");
+      await seedDatabase();
+    } else {
+      console.log("RESET_DATABASE is not set. Skipping database seeding.");
+    }
+
+    // Start server
+    app.listen(port, () => {
+      console.log(`Server running on http://localhost:${port}`);
+    });
+  } catch (error) {
+    console.error("Failed to start server:", error.message);
+    process.exit(1); // Exit the process on failure
   }
-})
+})();
 
-// Start defining your routes here
-app.get('/', (req, res) => {
-  //documentation
-  //Express List Endpoints
-  const endpoints = listEndpoints(app);
-
-  res.send({
-    message: "This API returns info of users and trips",
-    endpoints: endpoints
-  });
-});
-
-
-// Start the server
-app.listen(port, () => {
-  console.log(`Server running on http://localhost:${port}`)
-});
+// Simplify server.js to handle only the server initialization and database connection
